+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * ![Feature][badge-feature] Deployment is now more customizable and thus not as tied to Travis CI as before. ([#1147][github-1147])
 
+* ![Feature][badge-feature] Documenter now has builtin support for deploying from GitHub Actions. ([#1144][github-1144])
+
 * ![Enhancement][badge-enhancement] In the PDF/LaTeX output, images that are wider than the text are now being scaled down to text width automatically. The PDF builds now require the [adjustbox](https://ctan.org/pkg/adjustbox) LaTeX package to be available. ([#1137][github-1137])
 
 ## Version `v0.23.3`
@@ -414,6 +416,7 @@
 [github-1082]: https://github.com/JuliaDocs/Documenter.jl/pull/1082
 [github-1115]: https://github.com/JuliaDocs/Documenter.jl/pull/1115
 [github-1137]: https://github.com/JuliaDocs/Documenter.jl/pull/1137
+[github-1144]: https://github.com/JuliaDocs/Documenter.jl/pull/1144
 [github-1147]: https://github.com/JuliaDocs/Documenter.jl/pull/1147
 
 [documenterlatex]: https://github.com/JuliaDocs/DocumenterLaTeX.jl